requirements.txt
black
numpy
openai
pdfplumber
pymupdf
python-docx
ruff
sendgrid
streamlit
tiktoken
docx
pandas
newspaper3k
openpyxl